delay minutes to be long, same as millis and seconds quitexception, newly created rename system config file as Konfig.properties, and extra for KonfigExtra rename AppBase to App rename to JanitorMD5 using md5 password, and JanitorSIRAP sirap password give userr a chance to quit in the middle of inputting password and saying "fxxking tired of this game" timeout.minutes, password.encrypted go to user config

+++ sirap-basic/src/main/java/com/sirap/basic/component/MexedTimer.java
@@ -10,7 +10,7 @@ public abstract class MexedTimer {
 	private long delayMilliSeconds = -1;
 	private long periodMilliSeconds = -1;
     
-	protected void setDelayMinutes(double minutes) {
+	protected void setDelayMinutes(long minutes) {
 		this.delayMilliSeconds = (long)(minutes * Konstants.MILLI_PER_MINUTE);
 	}
